# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dutchkem Voice Agent
  version: 3.0.0
paths:
...
  /proctoring/register-face:
    post:
      tags:
      - proctoring
      summary: Register Face
      description: Register a face profile for a user.
      operationId: register_face_proctoring_register_face_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaceProfileCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaceProfileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /proctoring/verify:
    post:
      tags:
      - proctoring
      summary: Verify Face
      description: Verify if an image matches a registered face.
      operationId: verify_face_proctoring_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaceVerifyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaceVerifyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /voice/clone:
    post:
      tags:
      - voice
      summary: Clone Voice
      operationId: clone_voice_voice_clone_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_clone_voice_voice_clone_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloneResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /voice/synthesize:
    post:
      tags:
      - voice
      summary: Synthesize Voice
      operationId: synthesize_voice_voice_synthesize_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SynthesizeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SynthesizeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /voice/profiles:
    get:
      tags:
      - voice
      summary: List Profiles
      operationId: list_profiles_voice_profiles_get
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProfilesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /voice/tts:
    post:
      tags:
      - voice
      summary: Text To Speech
      description: Convert text to speech with accent/dialect support.
      operationId: text_to_speech_voice_tts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TTSRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TTSResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /voice/tts/stream:
    post:
      tags:
      - voice
      summary: Text To Speech Stream
      description: Stream text to speech audio chunks.
      operationId: text_to_speech_stream_voice_tts_stream_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TTSStreamRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /interview/start:
    post:
      tags:
      - interview
      summary: Start Session
      description: Start a new interview session.
      operationId: start_session_interview_start_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterviewSessionCreate'
              default:
                user_profile: {}
                company_context: {}
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewSessionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /interview/{session_id}/stop:
    post:
      tags:
      - interview
      summary: Stop Session
      description: Stop an interview session.
      operationId: stop_session_interview__session_id__stop_post
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewSessionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /interview/question:
    post:
      tags:
      - interview
      summary: Submit Question
      description: Submit a text question to the interview engine.
      operationId: submit_question_interview_question_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterviewQuestionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewAnswerResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /interview/audio:
    post:
      tags:
      - interview
      summary: Process Audio
      description: Process audio input through the interview engine.
      operationId: process_audio_interview_audio_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterviewAudioRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewAudioResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /interview/{session_id}/stats:
    get:
      tags:
      - interview
      summary: Get Session Stats
      description: Get stats for an active session.
      operationId: get_session_stats_interview__session_id__stats_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_clone_voice_voice_clone_post:
      properties:
        audio:
          type: string
          contentMediaType: application/octet-stream
          title: Audio
        user_id:
          type: string
          title: User Id
        name:
          type: string
          title: Name
          default: ''
      type: object
      required:
      - audio
      - user_id
      title: Body_clone_voice_voice_clone_post
...
    CloneResponse:
      properties:
        profile_id:
          type: string
          title: Profile Id
      type: object
      required:
      - profile_id
      title: CloneResponse
...
    FaceProfileCreate:
      properties:
        user_id:
          type: string
          title: User Id
        photo_path:
          type: string
          title: Photo Path
      type: object
      required:
      - user_id
      - photo_path
      title: FaceProfileCreate
    FaceProfileResponse:
      properties:
        user_id:
          type: string
          title: User Id
        message:
          type: string
          title: Message
        embedding_stored:
          type: boolean
          title: Embedding Stored
      type: object
      required:
      - user_id
      - message
      - embedding_stored
      title: FaceProfileResponse
    FaceVerifyRequest:
      properties:
        user_id:
          type: string
          title: User Id
        image_data:
          type: string
          title: Image Data
      type: object
      required:
      - user_id
      - image_data
      title: FaceVerifyRequest
    FaceVerifyResponse:
      properties:
        verified:
          type: boolean
          title: Verified
        confidence:
          type: number
          title: Confidence
      type: object
      required:
      - verified
      - confidence
      title: FaceVerifyResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InterviewAnswerResponse:
      properties:
        text:
          type: string
          title: Text
        confidence:
          type: number
          title: Confidence
        agent_type:
          type: string
          title: Agent Type
        follow_ups:
          items:
            type: string
          type: array
          title: Follow Ups
          default: []
        avatar_data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Avatar Data
      type: object
      required:
      - text
      - confidence
      - agent_type
      title: InterviewAnswerResponse
    InterviewAudioRequest:
      properties:
        session_id:
          type: string
          title: Session Id
        audio_base64:
          type: string
          title: Audio Base64
      type: object
      required:
      - session_id
      - audio_base64
      title: InterviewAudioRequest
    InterviewAudioResponse:
      properties:
        action:
          type: string
          title: Action
        transcript:
          anyOf:
          - type: string
          - type: 'null'
          title: Transcript
        answer:
          anyOf:
          - type: string
          - type: 'null'
          title: Answer
        confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence
        avatar:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Avatar
      type: object
      required:
      - action
      title: InterviewAudioResponse
    InterviewQuestionRequest:
      properties:
        session_id:
          type: string
          title: Session Id
        text:
          type: string
          title: Text
        category:
          type: string
          title: Category
          default: general
        difficulty:
          type: string
          title: Difficulty
          default: medium
      type: object
      required:
      - session_id
      - text
      title: InterviewQuestionRequest
    InterviewSessionCreate:
      properties:
        user_profile:
          additionalProperties: true
          type: object
          title: User Profile
          default: {}
        company_context:
          additionalProperties: true
          type: object
          title: Company Context
          default: {}
      type: object
      title: InterviewSessionCreate
    InterviewSessionResponse:
      properties:
        id:
          type: string
          title: Id
        status:
          type: string
          title: Status
        started_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Started At
        ended_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Ended At
      type: object
      required:
      - id
      - status
      title: InterviewSessionResponse
    ListProfilesResponse:
      properties:
        profiles:
          items:
            $ref: '#/components/schemas/VoiceProfileResponse'
          type: array
          title: Profiles
      type: object
      required:
      - profiles
      title: ListProfilesResponse
...
    SynthesizeRequest:
      properties:
        text:
          type: string
          title: Text
        profile_id:
          type: string
          title: Profile Id
        exaggeration:
          type: number
          title: Exaggeration
          default: 0.5
      type: object
      required:
      - text
      - profile_id
      title: SynthesizeRequest
    SynthesizeResponse:
      properties:
        audio_base64:
          type: string
          title: Audio Base64
        duration_ms:
          type: integer
          title: Duration Ms
          default: 0
      type: object
      required:
      - audio_base64
      title: SynthesizeResponse
    TTSRequest:
      properties:
        text:
          type: string
          maxLength: 5000
          minLength: 1
          title: Text
        voice:
          type: string
          title: Voice
          description: Voice name or accent (nigerian, ghanaian, uk, us, japanese)
          default: af_heart
      type: object
      required:
      - text
      title: TTSRequest
    TTSResponse:
      properties:
        audio_base64:
          type: string
          title: Audio Base64
        duration_ms:
          type: integer
          title: Duration Ms
          default: 0
      type: object
      required:
      - audio_base64
      title: TTSResponse
    TTSStreamRequest:
      properties:
        text:
          type: string
          maxLength: 5000
          minLength: 1
          title: Text
        voice:
          type: string
          title: Voice
          default: af_heart
      type: object
      required:
      - text
      title: TTSStreamRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VoiceProfileResponse:
      properties:
        profile_id:
          type: string
          title: Profile Id
        user_id:
          type: string
          title: User Id
        name:
          type: string
          title: Name
          default: ''
        reference_audio_path:
          type: string
          title: Reference Audio Path
          default: ''
        embedding:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Embedding
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - user_id
      title: VoiceProfileResponse
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer